refactor: separate analyticsProvider module
