+++ src/analyticsProvider.ts
@@ -1,0 +1,7 @@
+import Assignment from './assignment';
+
+export interface AnalyticsProvider {
+  trackAssignment(visitorId: string, assignment: Assignment, callback: (value: boolean) => void): void;
+  identify(visitorId: string): void;
+  alias(visitorId: string): void;
+}
